int is 32 bits. long is not. This is a 64 bit compatibility issue.
git-svn-id: http://svn.osgeo.org/mapguide/trunk@1166 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/MdfModel/ConvertUTF.h
@@ -90,7 +90,7 @@
 #ifndef CONVERT_UTF_H_
 #define CONVERT_UTF_H_
 
-typedef unsigned long   UTF32;  /* at least 32 bits */
+typedef unsigned int   UTF32;  /* at least 32 bits */
 typedef unsigned short  UTF16;  /* at least 16 bits */
 typedef unsigned char   UTF8;   /* typically 8 bits */
 typedef unsigned char   Boolean; /* 0 or 1 */
